requirements.txt
h5py==2.9.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.6
jupyterlab-server==0.2.0
matplotlib==3.1.0
notebook==5.7.8
numpy==1.16.4
opencv-python==3.4.1.15
pandas==0.24.2
Pillow==7.1.2
Pillow-SIMD==5.3.0.post1
scikit-image==0.17.2
scikit-learn==0.21.2
scipy==1.2.1
torch==1.4.0
torchtext==0.4.0
torchvision==0.5.0